+++ src/main/java/com/strathub2/backend2/repository/TimeRepository.java
@@ -5,6 +5,6 @@ import com.strathub2.backend2.entity.Time;
 
 import java.util.Optional;
 
-public interface TimeRepository extends JpaRepository<Time, Long>{
+public interface TimeRepository extends JpaRepository<Time, Integer>{
     Optional<Time> findByNome(String nome);
 }
