+++ src/main/java/com/strathub2/backend2/repository/UsuarioRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+
 
     Optional<Usuario> findByEmail(String email);
     Optional<Usuario> findById(Integer id);
